Minor refactoring
Moved non-views into their own files
TLLog views are now decoupled from the BP structure and self-contained
